# docs/api_spec.yaml
openapi: 3.0.3
info:
  title: My Jeja API
  version: 1.1.0
  description: |
    MVP API for My Jeja.
    Scope: math-only, levels 1-9, streaming teaching chat, secure exam grading.
servers:
  - url: https://your-service.up.railway.app/api/v1
...
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Register user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '429':
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      tags: [Auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '429':
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/me:
    get:
      tags: [Auth]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
  /auth/google:
    get:
      tags: [Auth]
      summary: Start Google OAuth redirect flow (backend-managed)
      parameters:
        - in: query
          name: redirect_uri
          required: false
          schema:
            type: string
      responses:
        '302':
          description: Redirect to Google OAuth consent screen
  /auth/google/login:
    get:
      tags: [Auth]
      summary: Start Google OAuth redirect flow (frontend-compatible endpoint)
      parameters:
        - in: query
          name: redirecturi
          required: false
          schema:
            type: string
      responses:
        '302':
          description: Redirect to Google OAuth consent screen
    post:
      tags: [Auth]
      summary: Login/signup with Google id_token (mobile/web GIS)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoogleLoginRequest'
      responses:
        '200':
          description: Access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /auth/google/callback:
    get:
      tags: [Auth]
      summary: Google OAuth callback endpoint (backend exchanges code and redirects with JWT)
      parameters:
        - in: query
          name: code
          required: true
          schema:
            type: string
        - in: query
          name: state
          required: true
          schema:
            type: string
      responses:
        '302':
          description: Redirect back to app/web with `access_token` in query string
...
  /auth/google/code:
    post:
      tags: [Auth]
      summary: Exchange Google auth code and login/signup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoogleCodeLoginRequest'
      responses:
        '200':
          description: Access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
...
  /persona:
    post:
      tags: [Persona]
      summary: Create persona (one per user in MVP)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePersonaRequest'
      responses:
        '200':
          description: Persona created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonaResponse'
        '409':
          description: Persona already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /persona/me:
    get:
      tags: [Persona]
      summary: Get current user's persona
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Persona summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonaResponse'
    patch:
      tags: [Persona]
      summary: Update persona profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePersonaRequest'
      responses:
        '200':
          description: Persona updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonaResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    GoogleLoginRequest:
      type: object
      required: [id_token]
      properties:
        id_token: { type: string }
    GoogleCodeLoginRequest:
      type: object
      required: [code, redirect_uri]
      properties:
        code: { type: string }
        redirect_uri: { type: string }
...
    AuthResponse:
      type: object
      properties:
        access_token: { type: string }
        token_type: { type: string, example: bearer }
    MeResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        created_at: { type: string, format: date-time }
    CreatePersonaRequest:
      type: object
      required: [name, personality]
      properties:
        name: { type: string }
        subject_id: { type: string, format: uuid, nullable: true }
        personality:
          type: string
          enum: [curious, careful, clumsy, perfectionist, steady]
    UpdatePersonaRequest:
      type: object
      properties:
        name: { type: string }
        personality:
          type: string
          enum: [curious, careful, clumsy, perfectionist, steady]
    PersonaResponse:
      type: object
      properties:
        persona_id: { type: string, format: uuid }
        name: { type: string }
        personality:
          type: string
          enum: [curious, careful, clumsy, perfectionist, steady]
        subject_id: { type: string, format: uuid, nullable: true }
        current_stage_id: { type: string, format: uuid, nullable: true }
        current_level: { type: integer, minimum: 1, maximum: 9 }
        placement_done: { type: boolean }
...
    ErrorResponse:
      type: object
      properties:
        detail: { type: string }
        request_id: { type: string }